Implement history step tracers
FEATURE: Adds support for tracers, which allow you to notice when a
step is undone or redone. The package exports a
[`tracers`](##history.tracers) key, which is used to optionally
associate an array of [`Tracer`](##history.Tracer) objects with a
transaction. When undoing or redoing, the tracers associated with the
undone or redone steps are added back to the newly generated
transaction.